requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-modeltranslation==0.11rc1
django-solo==1.1.1
django-toolbelt==0.0.1
gunicorn==19.4.1
html2text==2015.11.4
psutil==3.3.0
psycopg2==2.6.1
static3==0.6.1
